+++ Reflection/src/utopia/reflection/component/drawing/immutable/TextDrawContext.scala
@@ -18,10 +18,24 @@ import utopia.reflection.text.Font
 case class TextDrawContext(font: Font, color: Color = Color.textBlack, alignment: Alignment = Alignment.Left,
 						   insets: StackInsets = StackInsets.any, betweenLinesMargin: Double = 0.0)
 {
+	// OTHER	------------------------------------
+	
 	/**
 	  * @param other Another text draw context
 	  * @return Whether this context has different size-affecting qualities than the other context
 	  */
 	def hasSameDimensionsAs(other: TextDrawContext) = font == other.font && insets == other.insets &&
 		betweenLinesMargin == other.betweenLinesMargin
+	
+	/**
+	  * @param f A mapping function for font
+	  * @return A copy of this context with mapped font
+	  */
+	def mapFont(f: Font => Font) = copy(font = f(font))
+	
+	/**
+	  * @param f A mapping function for text color
+	  * @return A copy of this context with mapped text color
+	  */
+	def mapColor(f: Color => Color) = copy(color = f(color))
 }